Added JWT support with passport and jsonwebtoken

+++ controllers/api/v1/post_api.js
@@ -19,14 +19,20 @@ module.exports.index = async (req, res) =>{
 module.exports.destroy = async (req,res) => {
     try {
         let post = await Post.findById(req.params.id)
-        // if (post.user == req.user.id){
-            //Delete the post
+        if (post.user == req.user.id){
+            // Delete the post
             post.remove();
             await Comment.deleteMany({post: req.params.id})
             return res.status(200).json({
                 message: "Post deleted successfully",
                 success:true
             });
+        } else {
+            return res.status(401).json({
+                message:"You cannot delete this post",
+                success: false
+            })
+        }
     } catch (error) {
         return res.status(500).json({
             message: "Internal server error",
